+++ libs/langchain-community/src/embeddings/ibm.ts
@@ -5,7 +5,7 @@ import {
 } from "@ibm-cloud/watsonx-ai/dist/watsonx-ai-ml/vml_v1.js";
 import { WatsonXAI } from "@ibm-cloud/watsonx-ai";
 import { AsyncCaller } from "@langchain/core/utils/async_caller";
-import { WatsonxAuth, WatsonxParams } from "../types/watsonx_ai.js";
+import { WatsonxAuth, WatsonxParams } from "../types/ibm.js";
 import { authenticateAndSetInstance } from "../utils/ibm.js";
 
 export interface WatsonxEmbeddingsParams
