+++ libs/langchain-community/src/llms/tests/ibm.int.test.ts
@@ -2,7 +2,7 @@
 import { CallbackManager } from "@langchain/core/callbacks/manager";
 import { LLMResult } from "@langchain/core/outputs";
 import { StringPromptValue } from "@langchain/core/prompt_values";
-import { TokenUsage } from "../../types/watsonx_ai.js";
+import { TokenUsage } from "../../types/ibm.js";
 import { WatsonxLLM, WatsonxInputLLM } from "../ibm.js";
 
 const originalBackground = process.env.LANGCHAIN_CALLBACKS_BACKGROUND;
